refactor: split datasource.go into focused files by concern
The 1712-line datasource.go and 3710-line datasource_test.go had 4
distinct concerns that rarely need editing together. Split them into:

- datasource.go: struct, auth, JWT, URL building, CheckHealth (~330 lines)
- query.go: QueryData, frame/field transforms, type conversion (~410 lines)
- cubeclient.go: HTTP client, continue-wait protocol, metadata fetch (~250 lines)
- resources.go: CallResource router + all handle*/fetch* functions (~770 lines)

Test files mirror the split, with shared helpers in helpers_test.go.
No behavioral changes — pure structural refactoring.

Made-with: Cursor
