+++ pkg/plugin/datasource.go
@@ -3,12 +3,10 @@ package plugin
 import (
 	"context"
 	"encoding/json"
-	"errors"
 	"fmt"
 	"io"
 	"net/http"
 	"net/url"
-	"strconv"
 	"strings"
 	"sync"
 	"time"
@@ -17,8 +15,6 @@ import (
 	"github.com/grafana/cube/pkg/models"
 	"github.com/grafana/grafana-plugin-sdk-go/backend"
 	"github.com/grafana/grafana-plugin-sdk-go/backend/instancemgmt"
-	"github.com/grafana/grafana-plugin-sdk-go/data"
-	"github.com/grafana/grafana-plugin-sdk-go/data/framestruct"
 )
 
 // Make sure Datasource implements required interfaces. This is important to do
@@ -249,137 +245,6 @@ func (d *Datasource) buildAPIURL(pluginContext backend.PluginContext, endpoint s
 	}, nil
 }
 
-// QueryData handles multiple queries and returns multiple responses.
-// req contains the queries []DataQuery (where each query contains RefID as a unique identifier).
-// The QueryDataResponse contains a map of RefID to the response for each query, and each response
-// contains Frames ([]*Frame).
-func (d *Datasource) QueryData(ctx context.Context, req *backend.QueryDataRequest) (*backend.QueryDataResponse, error) {
-	// create response struct
-	response := backend.NewQueryDataResponse()
-
-	// loop over queries and execute them individually.
-	for _, q := range req.Queries {
-		res := d.query(ctx, req.PluginContext, q)
-
-		// save the response in a hashmap
-		// based on with RefID as identifier
-		response.Responses[q.RefID] = res
-	}
-
-	return response, nil
-}
-
-func (d *Datasource) query(ctx context.Context, pCtx backend.PluginContext, query backend.DataQuery) backend.DataResponse {
-	var response backend.DataResponse
-
-	// Ensure query JSON is provided
-	if len(query.JSON) == 0 {
-		return backend.ErrDataResponse(backend.StatusBadRequest, "Query JSON is required")
-	}
-
-	// Debug: Log the raw JSON to see what we're actually trying to unmarshal
-	backend.Logger.Debug("Raw query JSON", "rawJSON", string(query.JSON))
-
-	// Parse the query JSON into CubeQuery struct
-	var cubeQuery CubeQuery
-	if err := json.Unmarshal(query.JSON, &cubeQuery); err != nil {
-		return backend.ErrDataResponse(backend.StatusBadRequest, fmt.Sprintf("Invalid query JSON: %v", err))
-	}
-
-	backend.Logger.Debug("Parsed cube query", "measures", cubeQuery.Measures, "dimensions", cubeQuery.Dimensions, "timeDimensions", cubeQuery.TimeDimensions)
-
-	// Additional debugging: If arrays are empty, let's see the full JSON structure
-	if len(cubeQuery.Measures) == 0 && len(cubeQuery.Dimensions) == 0 {
-		var genericJSON map[string]interface{}
-		if err := json.Unmarshal(query.JSON, &genericJSON); err == nil {
-			backend.Logger.Debug("Full JSON structure", "structure", genericJSON)
-		}
-	}
-
-	// Build the Cube API query JSON (only include the Cube-specific fields)
-	cubeAPIQuery := map[string]interface{}{}
-	if len(cubeQuery.Dimensions) > 0 {
-		cubeAPIQuery["dimensions"] = cubeQuery.Dimensions
-	}
-	if len(cubeQuery.Measures) > 0 {
-		cubeAPIQuery["measures"] = cubeQuery.Measures
-	}
-	if len(cubeQuery.TimeDimensions) > 0 {
-		cubeAPIQuery["timeDimensions"] = cubeQuery.TimeDimensions
-	}
-	if len(cubeQuery.Filters) > 0 {
-		cubeAPIQuery["filters"] = cubeQuery.Filters
-	}
-	if cubeQuery.Order != nil {
-		cubeAPIQuery["order"] = cubeQuery.Order
-	}
-	if cubeQuery.Limit != nil {
-		cubeAPIQuery["limit"] = cubeQuery.Limit
-	}
-
-	cubeAPIQueryJSON, err := json.Marshal(cubeAPIQuery)
-	if err != nil {
-		return backend.ErrDataResponse(backend.StatusBadRequest, fmt.Sprintf("Failed to marshal Cube query: %v", err))
-	}
-
-	// Build API URL and load configuration
-	apiReq, err := d.buildAPIURL(pCtx, "load")
-	if err != nil {
-		return backend.ErrDataResponse(backend.StatusBadRequest, err.Error())
-	}
-
-	// Add query parameter
-	u, err := url.Parse(apiReq.URL.String())
-	if err != nil {
-		return backend.ErrDataResponse(backend.StatusBadRequest, fmt.Sprintf("Failed to parse API URL: %v", err))
-	}
-
-	params := url.Values{}
-	params.Add("query", string(cubeAPIQueryJSON))
-	u.RawQuery = params.Encode()
-
-	// Debug: Log what we're sending to the API
-	backend.Logger.Debug("Making API request", "url", u.String(), "cubeQuery", string(cubeAPIQueryJSON))
-
-	// Use shared helper to make the request with "Continue wait" polling
-	body, err := d.doCubeLoadRequest(ctx, u.String(), apiReq.Config)
-	if err != nil {
-		backend.Logger.Error("Failed to fetch data from Cube API", "error", err, "url", u.String())
-		return backend.ErrDataResponse(backend.StatusBadRequest, err.Error())
-	}
-
-	// Parse the API response
-	var apiResponse CubeAPIResponse
-	err = json.Unmarshal(body, &apiResponse)
-	if err != nil {
-		return backend.ErrDataResponse(backend.StatusBadRequest, fmt.Sprintf("Failed to parse API response: %v", err))
-	}
-
-	// Convert string values to numbers based on type annotations
-	convertedData := d.convertDataTypes(apiResponse.Data, apiResponse.Annotation)
-
-	// Create DataFrame using framestruct utility
-	frame, err := framestruct.ToDataFrame("response", convertedData)
-	if err != nil {
-		return backend.ErrDataResponse(backend.StatusBadRequest, fmt.Sprintf("Failed to convert data to frame: %v", err))
-	}
-
-	// Reorder fields according to query specification (dimensions first, then measures)
-	// Also adds missing fields (e.g., columns with all null values) as nullable fields
-	frame = d.reorderFrameFields(frame, cubeQuery, apiResponse.Annotation, len(apiResponse.Data))
-
-	// Mark dimension fields as filterable to enable AdHoc filter buttons
-	d.markFieldsAsFilterable(frame, cubeQuery)
-
-	// Convert time dimension strings to proper time.Time values for better UI display
-	d.convertTimeDimensions(frame, apiResponse.Annotation)
-
-	// add the frames to the response.
-	response.Frames = append(response.Frames, frame)
-
-	return response
-}
-
 // CheckHealth handles health checks sent from Grafana to the plugin.
 // The main use case for these health checks is the test button on the
 // datasource configuration page which allows users to verify that
@@ -461,1252 +326,3 @@ func (d *Datasource) CheckHealth(ctx context.Context, req *backend.CheckHealthRe
 	}, nil
 }
 
-// reorderFrameFields reorders the fields of a DataFrame according to the query specification.
-// It also adds missing fields (e.g., columns with all null values) as nullable fields.
-// The dataRowCount parameter is needed when all columns have null values (frame has no fields).
-func (d *Datasource) reorderFrameFields(frame *data.Frame, query CubeQuery, annotation CubeAnnotation, dataRowCount int) *data.Frame {
-	// Create a new frame with the reordered fields
-	newFrame := data.NewFrame(frame.Name)
-
-	// Create a map to track the new positions of the fields
-	fieldPositions := make(map[string]int)
-
-	// Populate the field positions map
-	for i, field := range frame.Fields {
-		fieldPositions[field.Name] = i
-	}
-
-	// Determine row count from existing frame (needed for creating null-filled fields)
-	// Fall back to dataRowCount when frame has no fields (all columns have null values)
-	rowCount := dataRowCount
-	if len(frame.Fields) > 0 {
-		rowCount = frame.Fields[0].Len()
-	}
-
-	// Reorder the fields according to the query specification
-	// If a field doesn't exist (all null values), create it as a nullable field
-	for _, fieldName := range query.Dimensions {
-		if pos, exists := fieldPositions[fieldName]; exists {
-			newFrame.Fields = append(newFrame.Fields, frame.Fields[pos])
-		} else {
-			// Field missing (all null values) - create a nullable field
-			newFrame.Fields = append(newFrame.Fields, d.createNullField(fieldName, rowCount, annotation))
-		}
-	}
-
-	for _, fieldName := range query.Measures {
-		if pos, exists := fieldPositions[fieldName]; exists {
-			newFrame.Fields = append(newFrame.Fields, frame.Fields[pos])
-		} else {
-			// Field missing (all null values) - create a nullable field
-			newFrame.Fields = append(newFrame.Fields, d.createNullField(fieldName, rowCount, annotation))
-		}
-	}
-
-	return newFrame
-}
-
-// createNullField creates a nullable field with nil values for columns that were omitted
-// from the Cube API response (because all values were null).
-func (d *Datasource) createNullField(fieldName string, rowCount int, annotation CubeAnnotation) *data.Field {
-	// Determine the field type from annotation
-	// Check all annotation maps: Dimensions, Measures, and TimeDimensions
-	fieldType := "string" // default
-	if info, ok := annotation.Dimensions[fieldName]; ok {
-		fieldType = info.Type
-	} else if info, ok := annotation.Measures[fieldName]; ok {
-		fieldType = info.Type
-	} else if info, ok := annotation.TimeDimensions[fieldName]; ok {
-		fieldType = info.Type
-	}
-
-	// Create a nullable field with nil values based on type
-	switch fieldType {
-	case "number":
-		values := make([]*float64, rowCount)
-		return data.NewField(fieldName, nil, values)
-	case "time":
-		values := make([]*time.Time, rowCount)
-		return data.NewField(fieldName, nil, values)
-	case "boolean":
-		values := make([]*bool, rowCount)
-		return data.NewField(fieldName, nil, values)
-	default:
-		// Default to nullable string
-		values := make([]*string, rowCount)
-		return data.NewField(fieldName, nil, values)
-	}
-}
-
-// markFieldsAsFilterable marks dimension fields as filterable to enable AdHoc filter buttons
-func (d *Datasource) markFieldsAsFilterable(frame *data.Frame, query CubeQuery) {
-	// Mark dimension fields as filterable
-	for _, field := range frame.Fields {
-		for _, dimension := range query.Dimensions {
-			if field.Name == dimension {
-				if field.Config == nil {
-					field.Config = &data.FieldConfig{}
-				}
-				// Set Filterable to true to enable "Filter for value" buttons on table cell hover
-				field.Config.Filterable = &[]bool{true}[0] // Convert to *bool
-				break
-			}
-		}
-	}
-}
-
-// convertTimeDimensions converts time dimension string fields to proper time.Time values.
-// This enables proper time formatting and sorting in Grafana's UI.
-func (d *Datasource) convertTimeDimensions(frame *data.Frame, annotation CubeAnnotation) {
-	for i, field := range frame.Fields {
-		// Check if this is a time dimension field (from timeDimensions annotation)
-		if timeDimInfo, isTimeDim := annotation.TimeDimensions[field.Name]; isTimeDim {
-			if timeDimInfo.Type == "time" {
-				newField := d.convertTimeField(field)
-				if newField != nil {
-					frame.Fields[i] = newField
-				}
-				continue
-			}
-		}
-
-		// Also check regular dimensions that have type "time"
-		// (e.g., date fields used as regular dimensions without granularity)
-		if dimInfo, isDim := annotation.Dimensions[field.Name]; isDim {
-			if dimInfo.Type == "time" {
-				newField := d.convertTimeField(field)
-				if newField != nil {
-					frame.Fields[i] = newField
-				}
-			}
-		}
-	}
-}
-
-// convertTimeField converts string time values to proper time.Time values.
-// Returns nil if the field is not a string type (no conversion needed).
-func (d *Datasource) convertTimeField(field *data.Field) *data.Field {
-	if field.Type() != data.FieldTypeString && field.Type() != data.FieldTypeNullableString {
-		return nil // Already proper type or not convertible
-	}
-
-	// Create a new time vector with the same length
-	timeValues := make([]*time.Time, field.Len())
-
-	for i := 0; i < field.Len(); i++ {
-		val := field.At(i)
-		if val == nil {
-			continue
-		}
-
-		var timeStr string
-		switch v := val.(type) {
-		case string:
-			timeStr = v
-		case *string:
-			if v != nil {
-				timeStr = *v
-			}
-		default:
-			continue
-		}
-
-		if timeStr == "" {
-			continue
-		}
-
-		// Try parsing common time formats used by Cube
-		if t, err := time.Parse(time.RFC3339, timeStr); err == nil {
-			timeValues[i] = &t
-		} else if t, err := time.Parse("2006-01-02T15:04:05.000Z", timeStr); err == nil {
-			timeValues[i] = &t
-		} else if t, err := time.Parse("2006-01-02T15:04:05.000", timeStr); err == nil {
-			timeValues[i] = &t
-		} else if t, err := time.Parse("2006-01-02", timeStr); err == nil {
-			timeValues[i] = &t
-		}
-		// If parsing fails, keep as nil
-	}
-
-	// Create a new time field and copy values
-	newField := data.NewField(field.Name, field.Labels, timeValues)
-	newField.Config = field.Config
-
-	return newField
-}
-
-// convertDataTypes converts string values to numbers based on type annotations from Cube API
-func (d *Datasource) convertDataTypes(data []map[string]interface{}, annotation CubeAnnotation) []map[string]interface{} {
-	convertedData := make([]map[string]interface{}, len(data))
-
-	// Create a combined map of all field types for easy lookup
-	fieldTypes := make(map[string]string)
-	for fieldName, info := range annotation.Measures {
-		fieldTypes[fieldName] = info.Type
-	}
-	for fieldName, info := range annotation.Dimensions {
-		fieldTypes[fieldName] = info.Type
-	}
-	for fieldName, info := range annotation.Segments {
-		fieldTypes[fieldName] = info.Type
-	}
-	for fieldName, info := range annotation.TimeDimensions {
-		fieldTypes[fieldName] = info.Type
-	}
-
-	// Convert each row
-	for i, row := range data {
-		convertedRow := make(map[string]interface{})
-		for fieldName, value := range row {
-			if fieldTypes[fieldName] == "number" {
-				convertedRow[fieldName] = d.convertToNumber(value)
-			} else {
-				convertedRow[fieldName] = value
-			}
-		}
-		convertedData[i] = convertedRow
-	}
-
-	return convertedData
-}
-
-// convertToNumber attempts to convert a value to a number if it's a string representation of a number
-// Always return float64. Fields within Grafana DataFrame cannot have a mix of types
-func (d *Datasource) convertToNumber(value interface{}) interface{} {
-	switch v := value.(type) {
-	case string:
-		// Try to parse as float (handles both integers and decimals)
-		if floatVal, err := strconv.ParseFloat(v, 64); err == nil {
-			return floatVal
-		}
-		// If parsing fails, return the original string
-		return v
-	case int:
-		return float64(v)
-	case int8:
-		return float64(v)
-	case int16:
-		return float64(v)
-	case int32:
-		return float64(v)
-	case int64:
-		return float64(v)
-	case uint:
-		return float64(v)
-	case uint8:
-		return float64(v)
-	case uint16:
-		return float64(v)
-	case uint32:
-		return float64(v)
-	case uint64:
-		return float64(v)
-	case float32:
-		return float64(v)
-	case float64:
-		return v
-	default:
-		// For any other type, return as is
-		return v
-	}
-}
-
-// CubeAPIError represents a non-200 HTTP response from the Cube API.
-// It preserves the original status code and body so callers (e.g. handleTagValues)
-// can forward them to the frontend instead of collapsing everything to 500.
-type CubeAPIError struct {
-	StatusCode int
-	Body       []byte
-}
-
-func (e *CubeAPIError) Error() string {
-	return fmt.Sprintf("API request failed with status %d: %s", e.StatusCode, string(e.Body))
-}
-
-// doCubeLoadRequest makes a GET request to Cube's /v1/load endpoint, handling the
-// "Continue wait" polling protocol. Cube returns {"error": "Continue wait"} (HTTP 200)
-// when query results aren't cached yet (e.g. the upstream warehouse is still computing).
-// This method retries immediately until actual data arrives or the context is cancelled, matching the
-// behavior of the official @cubejs-client/core SDK.
-func (d *Datasource) doCubeLoadRequest(ctx context.Context, requestURL string, config *models.PluginSettings) ([]byte, error) {
-	pollStart := time.Now()
-	pollRetries := 0
-	var lastContinueWaitProgress continueWaitProgress
-	haveContinueWaitProgress := false
-	for {
-		req, err := http.NewRequestWithContext(ctx, "GET", requestURL, nil)
-		if err != nil {
-			return nil, fmt.Errorf("failed to create request: %w", err)
-		}
-
-		if err := d.addAuthHeaders(req, config); err != nil {
-			return nil, fmt.Errorf("failed to add auth headers: %w", err)
-		}
-		req.Header.Set("Content-Type", "application/json")
-
-		client := &http.Client{}
-		resp, err := client.Do(req)
-		if err != nil {
-			if errors.Is(err, context.DeadlineExceeded) {
-				elapsed := time.Since(pollStart).Round(time.Millisecond)
-				msg := fmt.Sprintf("request to Cube API timed out after %s (the upstream warehouse may still be computing)", elapsed)
-				if haveContinueWaitProgress && (lastContinueWaitProgress.Stage != "" || lastContinueWaitProgress.TimeElapsed > 0) {
-					msg = fmt.Sprintf("%s (stage: %s, Cube timeElapsed: %ds)", msg, lastContinueWaitProgress.Stage, int(lastContinueWaitProgress.TimeElapsed))
-				}
-				return nil, fmt.Errorf("%s", msg)
-			}
-			if errors.Is(err, context.Canceled) {
-				msg := "query cancelled while waiting for Cube to compute results"
-				if haveContinueWaitProgress && (lastContinueWaitProgress.Stage != "" || lastContinueWaitProgress.TimeElapsed > 0) {
-					msg = fmt.Sprintf("%s (stage: %s, Cube timeElapsed: %ds)", msg, lastContinueWaitProgress.Stage, int(lastContinueWaitProgress.TimeElapsed))
-				}
-				return nil, fmt.Errorf("%s", msg)
-			}
-			return nil, fmt.Errorf("failed to make API request: %w", err)
-		}
-
-		if resp.StatusCode != http.StatusOK {
-			errorBody, _ := io.ReadAll(resp.Body)
-			_ = resp.Body.Close()
-			return nil, &CubeAPIError{StatusCode: resp.StatusCode, Body: errorBody}
-		}
-
-		body, err := io.ReadAll(resp.Body)
-		_ = resp.Body.Close()
-		if err != nil {
-			return nil, fmt.Errorf("failed to read response body: %w", err)
-		}
-
-		if isContinueWait(body) {
-			// Parse progress info from the response for logging and error messages.
-			// Cube returns {"error": "Continue wait", "stage": "...", "timeElapsed": N}
-			progress := parseContinueWaitProgress(body)
-			lastContinueWaitProgress = progress
-			haveContinueWaitProgress = true
-
-			if pollRetries == 0 {
-				backend.Logger.Info("Cube query not yet ready, polling for results", "url", requestURL)
-			}
-			pollRetries++
-			backend.Logger.Debug("Cube returned 'Continue wait', polling again",
-				"url", requestURL, "attempt", pollRetries,
-				"stage", progress.Stage, "cubeTimeElapsed", progress.TimeElapsed)
-			select {
-			case <-ctx.Done():
-				var msg string
-				if errors.Is(ctx.Err(), context.DeadlineExceeded) {
-					msg = "Cube API request timed out while waiting for results to be computed"
-				} else {
-					msg = "query cancelled while waiting for Cube to compute results"
-				}
-				if progress.Stage != "" || progress.TimeElapsed > 0 {
-					msg = fmt.Sprintf("%s (stage: %s, Cube timeElapsed: %ds)", msg, progress.Stage, int(progress.TimeElapsed))
-				}
-				return nil, fmt.Errorf("%s", msg)
-			default:
-				continue
-			}
-		}
-
-		if pollRetries > 0 {
-			backend.Logger.Info("Cube query results ready after polling", "url", requestURL, "retries", pollRetries, "duration", time.Since(pollStart).Round(time.Millisecond))
-		}
-
-		return body, nil
-	}
-}
-
-// isContinueWait checks whether a Cube API response body is a "Continue wait"
-// polling response. Cube returns {"error": "Continue wait"} (HTTP 200) when
-// the query result is not yet ready (e.g. the upstream warehouse is still
-// computing). The caller is expected to retry until actual data arrives.
-func isContinueWait(body []byte) bool {
-	var probe struct {
-		Error string `json:"error"`
-	}
-	if err := json.Unmarshal(body, &probe); err != nil {
-		return false
-	}
-	return probe.Error == "Continue wait"
-}
-
-// continueWaitProgress holds progress information from a Cube "Continue wait" response.
-type continueWaitProgress struct {
-	Stage       string  `json:"stage"`
-	TimeElapsed float64 `json:"timeElapsed"`
-}
-
-// parseContinueWaitProgress extracts stage and timeElapsed from a Cube
-// "Continue wait" response body. Returns zero values if fields are missing.
-func parseContinueWaitProgress(body []byte) continueWaitProgress {
-	var progress continueWaitProgress
-	_ = json.Unmarshal(body, &progress)
-	return progress
-}
-
-// CubeAPIResponse represents the response structure from Cube API
-type CubeAPIResponse struct {
-	Data       []map[string]interface{} `json:"data"`
-	Annotation CubeAnnotation           `json:"annotation"`
-}
-
-// CubeAnnotation represents the type information from Cube API
-type CubeAnnotation struct {
-	Measures       map[string]CubeFieldInfo `json:"measures"`
-	Dimensions     map[string]CubeFieldInfo `json:"dimensions"`
-	Segments       map[string]CubeFieldInfo `json:"segments"`
-	TimeDimensions map[string]CubeFieldInfo `json:"timeDimensions"`
-}
-
-// CubeFieldInfo represents the metadata for a field
-type CubeFieldInfo struct {
-	Title      string `json:"title"`
-	ShortTitle string `json:"shortTitle"`
-	Type       string `json:"type"`
-}
-
-// CubeQuery represents the structure of a Cube query
-type CubeQuery struct {
-	RefID          string        `json:"refId"`
-	Measures       []string      `json:"measures"`
-	Dimensions     []string      `json:"dimensions"`
-	TimeDimensions []interface{} `json:"timeDimensions,omitempty"`
-	Filters        []interface{} `json:"filters,omitempty"`
-	Order          interface{}   `json:"order,omitempty"`
-	Limit          *int          `json:"limit,omitempty"`
-}
-
-// fetchCubeMetadata fetches metadata from Cube's /v1/meta endpoint
-func (d *Datasource) fetchCubeMetadata(ctx context.Context, pluginContext backend.PluginContext) (*CubeMetaResponse, error) {
-	// Build API URL and load configuration
-	apiReq, err := d.buildAPIURL(pluginContext, "meta")
-	if err != nil {
-		return nil, err
-	}
-
-	// Create HTTP request
-	req, err := http.NewRequestWithContext(ctx, "GET", apiReq.URL.String(), nil)
-	if err != nil {
-		return nil, fmt.Errorf("failed to create request: %w", err)
-	}
-
-	if err := d.addAuthHeaders(req, apiReq.Config); err != nil {
-		return nil, fmt.Errorf("failed to add auth headers: %w", err)
-	}
-	req.Header.Set("Content-Type", "application/json")
-
-	// Make the HTTP request
-	client := &http.Client{}
-	resp, err := client.Do(req)
-	if err != nil {
-		return nil, fmt.Errorf("failed to make API request: %w", err)
-	}
-	defer func() {
-		if err := resp.Body.Close(); err != nil {
-			backend.Logger.Warn("Failed to close response body", "error", err)
-		}
-	}()
-
-	// Check response status
-	if resp.StatusCode != http.StatusOK {
-		errorBody, _ := io.ReadAll(resp.Body)
-		return nil, fmt.Errorf("API request failed with status %d: %s", resp.StatusCode, string(errorBody))
-	}
-
-	// Read response body
-	body, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return nil, fmt.Errorf("failed to read response body: %w", err)
-	}
-
-	// Parse the API response
-	var metaResponse CubeMetaResponse
-	if err := json.Unmarshal(body, &metaResponse); err != nil {
-		return nil, fmt.Errorf("failed to parse API response: %w", err)
-	}
-
-	return &metaResponse, nil
-}
-
-// extractMetadataFromResponse extracts dimensions and measures from Cube metadata
-func (d *Datasource) extractMetadataFromResponse(metaResponse *CubeMetaResponse) MetadataResponse {
-	var dimensions []SelectOption
-	var measures []SelectOption
-
-	// Filter to only include views from the cubes array
-	var views []CubeMeta
-	var cubes []CubeMeta
-	for _, cube := range metaResponse.Cubes {
-		if cube.Type == "view" {
-			views = append(views, cube)
-		} else {
-			cubes = append(cubes, cube)
-		}
-	}
-
-	// Use views if available (they represent the curated data model), otherwise fall back to cubes
-	var itemsToProcess []CubeMeta
-	if len(views) > 0 {
-		itemsToProcess = views
-		backend.Logger.Debug("Using views for metadata", "viewCount", len(views))
-	} else {
-		itemsToProcess = cubes
-		backend.Logger.Debug("Using cubes for metadata (no views found)", "cubeCount", len(cubes))
-	}
-
-	// Track processed items to avoid duplicates
-	processedDimensions := make(map[string]bool)
-	processedMeasures := make(map[string]bool)
-
-	// Iterate through views/cubes and collect dimensions and measures
-	for _, item := range itemsToProcess {
-		// Collect dimensions
-		for _, dimension := range item.Dimensions {
-			if !processedDimensions[dimension.Name] {
-				dimensions = append(dimensions, SelectOption{
-					Label: dimension.Name,
-					Value: dimension.Name,
-					Type:  dimension.Type,
-				})
-				processedDimensions[dimension.Name] = true
-			}
-		}
-
-		// Collect measures
-		for _, measure := range item.Measures {
-			if !processedMeasures[measure.Name] {
-				measures = append(measures, SelectOption{
-					Label: measure.Name,
-					Value: measure.Name,
-					Type:  measure.Type,
-				})
-				processedMeasures[measure.Name] = true
-			}
-		}
-	}
-
-	backend.Logger.Debug("Extracted metadata", "dimensions", len(dimensions), "measures", len(measures))
-
-	return MetadataResponse{
-		Dimensions: dimensions,
-		Measures:   measures,
-	}
-}
-
-// CallResource handles resource calls for AdHoc filtering
-func (d *Datasource) CallResource(ctx context.Context, req *backend.CallResourceRequest, sender backend.CallResourceResponseSender) error {
-	switch req.Path {
-	case "tag-values":
-		return d.handleTagValues(ctx, req, sender)
-	case "sql":
-		return d.handleSQLCompilation(ctx, req, sender)
-	case "metadata":
-		return d.handleMetadata(ctx, req, sender)
-	case "model-files":
-		return d.handleModelFiles(ctx, req, sender)
-	case "db-schema":
-		return d.handleDbSchema(ctx, req, sender)
-	case "generate-schema":
-		return d.handleGenerateSchema(ctx, req, sender)
-	default:
-		return sender.Send(&backend.CallResourceResponse{
-			Status: 404,
-			Body:   []byte(`{"error": "not found"}`),
-		})
-	}
-}
-
-func jsonErrorResponse(status int, err error) *backend.CallResourceResponse {
-	body, marshalErr := json.Marshal(map[string]string{"error": err.Error()})
-	if marshalErr != nil {
-		body = []byte(`{"error":"internal server error"}`)
-	}
-
-	return &backend.CallResourceResponse{
-		Status: status,
-		Body:   body,
-		Headers: map[string][]string{
-			"Content-Type": {"application/json"},
-		},
-	}
-}
-
-// handleMetadata returns dimensions and measures for the query builder
-func (d *Datasource) handleMetadata(ctx context.Context, req *backend.CallResourceRequest, sender backend.CallResourceResponseSender) error {
-	// Fetch metadata from Cube API
-	metaResponse, err := d.fetchCubeMetadata(ctx, req.PluginContext)
-	if err != nil {
-		backend.Logger.Error("Failed to fetch cube metadata", "error", err)
-		return sender.Send(jsonErrorResponse(500, errors.New("failed to fetch metadata from Cube API")))
-	}
-
-	// Extract dimensions and measures from metadata
-	metadata := d.extractMetadataFromResponse(metaResponse)
-
-	// Marshal response
-	body, err := json.Marshal(metadata)
-	if err != nil {
-		backend.Logger.Error("Failed to marshal metadata response", "error", err)
-		return sender.Send(jsonErrorResponse(500, errors.New("failed to marshal response")))
-	}
-
-	return sender.Send(&backend.CallResourceResponse{
-		Status: 200,
-		Body:   body,
-		Headers: map[string][]string{
-			"Content-Type": {"application/json"},
-		},
-	})
-}
-
-// handleTagValues returns available tag values for a given tag key (dimension)
-// It queries the Cube /v1/load endpoint with just the dimension to get distinct values
-func (d *Datasource) handleTagValues(ctx context.Context, req *backend.CallResourceRequest, sender backend.CallResourceResponseSender) error {
-	// Parse the URL to get the key parameter
-	parsedURL, err := url.Parse(req.URL)
-	if err != nil {
-		return sender.Send(jsonErrorResponse(400, errors.New("invalid URL")))
-	}
-
-	key := parsedURL.Query().Get("key")
-	if key == "" {
-		return sender.Send(jsonErrorResponse(400, errors.New("key parameter is required")))
-	}
-
-	// Build a Cube query to get distinct values for this dimension
-	cubeQuery := map[string]interface{}{
-		"dimensions": []string{key},
-		"limit":      10000, // Limit for tag value suggestions
-	}
-
-	// Parse existing filters to scope the results (like Prometheus does)
-	filtersJSON := parsedURL.Query().Get("filters")
-	if filtersJSON != "" {
-		var scopingFilters []map[string]interface{}
-		if err := json.Unmarshal([]byte(filtersJSON), &scopingFilters); err != nil {
-			backend.Logger.Warn("Failed to parse scoping filters, ignoring", "error", err)
-		} else if len(scopingFilters) > 0 {
-			cubeQuery["filters"] = scopingFilters
-			backend.Logger.Debug("Scoping tag values with existing filters", "filters", scopingFilters)
-		}
-	}
-
-	cubeQueryJSON, err := json.Marshal(cubeQuery)
-	if err != nil {
-		return sender.Send(jsonErrorResponse(500, errors.New("failed to marshal query")))
-	}
-
-	// Build API URL
-	apiReq, err := d.buildAPIURL(req.PluginContext, "load")
-	if err != nil {
-		backend.Logger.Error("Failed to build API URL for tag values", "error", err)
-		return sender.Send(jsonErrorResponse(500, fmt.Errorf("failed to build API URL: %w", err)))
-	}
-
-	// Add query parameter
-	u, err := url.Parse(apiReq.URL.String())
-	if err != nil {
-		return sender.Send(jsonErrorResponse(500, errors.New("failed to parse API URL")))
-	}
-
-	params := url.Values{}
-	params.Add("query", string(cubeQueryJSON))
-	u.RawQuery = params.Encode()
-
-	// Use shared helper to make the request with "Continue wait" polling
-	body, err := d.doCubeLoadRequest(ctx, u.String(), apiReq.Config)
-	if err != nil {
-		backend.Logger.Error("Failed to fetch tag values from Cube API", "error", err)
-		// If this is a Cube API error (non-200), forward the original status code and body
-		var cubeErr *CubeAPIError
-		if errors.As(err, &cubeErr) {
-			return sender.Send(&backend.CallResourceResponse{
-				Status: cubeErr.StatusCode,
-				Body:   cubeErr.Body,
-				Headers: map[string][]string{
-					"Content-Type": {"application/json"},
-				},
-			})
-		}
-		// For other errors (timeouts, network, etc.), return 500 with safely encoded JSON
-		return sender.Send(jsonErrorResponse(500, err))
-	}
-
-	// Parse the Cube API response
-	var apiResponse CubeAPIResponse
-	if err := json.Unmarshal(body, &apiResponse); err != nil {
-		backend.Logger.Error("Failed to parse Cube API response for tag values", "error", err, "body", string(body))
-		return sender.Send(jsonErrorResponse(500, errors.New("failed to parse API response")))
-	}
-
-	// Extract unique values from the response data
-	// Response format for Grafana: [{ "text": "value1" }, { "text": "value2" }]
-	tagValues := []TagValue{}
-	seen := make(map[string]bool)
-
-	for _, row := range apiResponse.Data {
-		if value, ok := row[key]; ok && value != nil {
-			// Convert value to string
-			var strValue string
-			switch v := value.(type) {
-			case string:
-				strValue = v
-			case float64:
-				strValue = fmt.Sprintf("%v", v)
-			case bool:
-				strValue = fmt.Sprintf("%v", v)
-			default:
-				strValue = fmt.Sprintf("%v", v)
-			}
-
-			// Only add unique values
-			if !seen[strValue] {
-				seen[strValue] = true
-				tagValues = append(tagValues, TagValue{Text: strValue})
-			}
-		}
-	}
-
-	// Marshal response
-	responseBody, err := json.Marshal(tagValues)
-	if err != nil {
-		return sender.Send(jsonErrorResponse(500, errors.New("failed to marshal response")))
-	}
-
-	return sender.Send(&backend.CallResourceResponse{
-		Status: 200,
-		Body:   responseBody,
-		Headers: map[string][]string{
-			"Content-Type": {"application/json"},
-		},
-	})
-}
-
-// handleSQLCompilation compiles a Cube query to SQL using Cube's /v1/sql endpoint
-func (d *Datasource) handleSQLCompilation(ctx context.Context, req *backend.CallResourceRequest, sender backend.CallResourceResponseSender) error {
-	// Parse the URL to get query parameters
-	parsedURL, err := url.Parse(req.URL)
-	if err != nil {
-		return sender.Send(jsonErrorResponse(400, errors.New("invalid URL")))
-	}
-
-	// Get the query from URL parameters
-	queryParam := parsedURL.Query().Get("query")
-	if queryParam == "" {
-		return sender.Send(jsonErrorResponse(400, errors.New("query parameter is required")))
-	}
-
-	// Validate that it's valid JSON
-	var cubeQuery CubeQuery
-	if err := json.Unmarshal([]byte(queryParam), &cubeQuery); err != nil {
-		return sender.Send(jsonErrorResponse(400, errors.New("invalid query JSON")))
-	}
-
-	// Fetch SQL from Cube API
-	sqlString, err := d.fetchCubeSQL(ctx, req.PluginContext, queryParam)
-	if err != nil {
-		backend.Logger.Error("Failed to fetch SQL from Cube", "error", err)
-		return sender.Send(jsonErrorResponse(500, err))
-	}
-
-	// Return the SQL string
-	sqlJSON := map[string]string{"sql": sqlString}
-	responseBody, err := json.Marshal(sqlJSON)
-	if err != nil {
-		backend.Logger.Error("Failed to marshal SQL response", "error", err)
-		return sender.Send(jsonErrorResponse(500, errors.New("failed to marshal response")))
-	}
-
-	return sender.Send(&backend.CallResourceResponse{
-		Status: 200,
-		Body:   responseBody,
-		Headers: map[string][]string{
-			"Content-Type": {"application/json"},
-		},
-	})
-}
-
-// fetchCubeSQL compiles a Cube query to SQL using Cube's /v1/sql endpoint
-func (d *Datasource) fetchCubeSQL(ctx context.Context, pluginContext backend.PluginContext, query string) (string, error) {
-	// Build API URL and load configuration
-	apiReq, err := d.buildAPIURL(pluginContext, "sql")
-	if err != nil {
-		return "", fmt.Errorf("failed to build API URL: %w", err)
-	}
-
-	// Add query parameter
-	u, err := url.Parse(apiReq.URL.String())
-	if err != nil {
-		return "", fmt.Errorf("failed to parse API URL: %w", err)
-	}
-
-	params := url.Values{}
-	params.Add("query", query)
-	u.RawQuery = params.Encode()
-
-	// Create HTTP request
-	req, err := http.NewRequestWithContext(ctx, "GET", u.String(), nil)
-	if err != nil {
-		return "", fmt.Errorf("failed to create request: %w", err)
-	}
-
-	// Add authentication headers
-	if err := d.addAuthHeaders(req, apiReq.Config); err != nil {
-		return "", fmt.Errorf("failed to add auth headers: %w", err)
-	}
-	req.Header.Set("Content-Type", "application/json")
-
-	// Make the HTTP request
-	client := &http.Client{}
-	resp, err := client.Do(req)
-	if err != nil {
-		return "", fmt.Errorf("failed to make API request: %w", err)
-	}
-	defer func() {
-		if err := resp.Body.Close(); err != nil {
-			backend.Logger.Warn("Failed to close response body", "error", err)
-		}
-	}()
-
-	// Read response body
-	body, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return "", fmt.Errorf("failed to read response body: %w", err)
-	}
-
-	// Check response status
-	if resp.StatusCode != http.StatusOK {
-		return "", fmt.Errorf("API request failed with status %d: %s", resp.StatusCode, string(body))
-	}
-
-	// Parse the SQL API response
-	var sqlResponse CubeSQLResponse
-	if err := json.Unmarshal(body, &sqlResponse); err != nil {
-		return "", fmt.Errorf("failed to parse API response: %w", err)
-	}
-
-	// Extract SQL string from nested structure: response.sql.sql[0]
-	if len(sqlResponse.SQL.SQL) == 0 {
-		return "", fmt.Errorf("SQL array is empty")
-	}
-
-	sql, ok := sqlResponse.SQL.SQL[0].(string)
-	if !ok {
-		return "", fmt.Errorf("SQL response is not a string")
-	}
-
-	return sql, nil
-}
-
-// CubeSQLResponse represents the response from Cube's /v1/sql endpoint
-type CubeSQLResponse struct {
-	SQL struct {
-		SQL []interface{} `json:"sql"` // [sqlString, parameters]
-	} `json:"sql"`
-}
-
-// CubeMetaResponse represents the response from Cube's /v1/meta endpoint
-type CubeMetaResponse struct {
-	Cubes []CubeMeta `json:"cubes"` // Contains both cubes and views, distinguished by the Type field
-}
-
-// CubeMeta represents metadata for a single cube or view
-type CubeMeta struct {
-	Name       string          `json:"name"`
-	Title      string          `json:"title"`
-	Type       string          `json:"type"` // "cube" or "view"
-	Dimensions []CubeDimension `json:"dimensions"`
-	Measures   []CubeMeasure   `json:"measures"`
-}
-
-// CubeDimension represents a dimension in a cube
-type CubeDimension struct {
-	Name       string `json:"name"`
-	Title      string `json:"title"`
-	Type       string `json:"type"`
-	ShortTitle string `json:"shortTitle"`
-}
-
-// CubeMeasure represents a measure in a cube
-type CubeMeasure struct {
-	Name       string `json:"name"`
-	Title      string `json:"title"`
-	Type       string `json:"type"`
-	ShortTitle string `json:"shortTitle"`
-}
-
-// TagKey represents a tag key for AdHoc filtering
-type TagKey struct {
-	Text  string `json:"text"`
-	Value string `json:"value"`
-}
-
-// TagValue represents a tag value for AdHoc filtering
-type TagValue struct {
-	Text string `json:"text"`
-}
-
-// MetadataResponse represents the response for the metadata endpoint
-type MetadataResponse struct {
-	Dimensions []SelectOption `json:"dimensions"`
-	Measures   []SelectOption `json:"measures"`
-}
-
-// SelectOption represents an option for select components
-type SelectOption struct {
-	Label string `json:"label"`
-	Value string `json:"value"`
-	Type  string `json:"type"`
-}
-
-// ModelFile represents a data model file from Cube
-type ModelFile struct {
-	FileName string `json:"fileName"`
-	Content  string `json:"content"`
-}
-
-// ModelFilesResponse represents the response for the model-files endpoint
-type ModelFilesResponse struct {
-	Files []ModelFile `json:"files"`
-}
-
-// handleModelFiles fetches data model files from the Cube API
-func (d *Datasource) handleModelFiles(ctx context.Context, req *backend.CallResourceRequest, sender backend.CallResourceResponseSender) error {
-	// Fetch model files from Cube API
-	modelFiles, err := d.fetchCubeModelFiles(ctx, req.PluginContext)
-	if err != nil {
-		backend.Logger.Error("Failed to fetch cube model files", "error", err)
-		return sender.Send(jsonErrorResponse(500, errors.New("failed to fetch model files from Cube API")))
-	}
-
-	// Marshal response
-	body, err := json.Marshal(modelFiles)
-	if err != nil {
-		backend.Logger.Error("Failed to marshal model files response", "error", err)
-		return sender.Send(jsonErrorResponse(500, errors.New("failed to marshal response")))
-	}
-
-	return sender.Send(&backend.CallResourceResponse{
-		Status: 200,
-		Body:   body,
-		Headers: map[string][]string{
-			"Content-Type": {"application/json"},
-		},
-	})
-}
-
-// fetchCubeModelFiles fetches model files from Cube's /playground/files endpoint
-func (d *Datasource) fetchCubeModelFiles(ctx context.Context, pluginContext backend.PluginContext) (*ModelFilesResponse, error) {
-	// Build base URL and load configuration
-	apiReq, err := d.buildAPIURL(pluginContext, "")
-	if err != nil {
-		return nil, err
-	}
-
-	// Get base URL with test override support
-	baseURL := apiReq.Config.CubeApiUrl
-	if d.BaseURL != "" {
-		// Override for testing
-		baseURL = d.BaseURL
-	}
-
-	// Construct playground files URL
-	baseURL = strings.TrimRight(baseURL, "/")
-	filesURL := baseURL + "/playground/files"
-
-	// Create HTTP request
-	req, err := http.NewRequestWithContext(ctx, "GET", filesURL, nil)
-	if err != nil {
-		return nil, fmt.Errorf("failed to create request: %w", err)
-	}
-
-	if err := d.addAuthHeaders(req, apiReq.Config); err != nil {
-		return nil, fmt.Errorf("failed to add auth headers: %w", err)
-	}
-	req.Header.Set("Content-Type", "application/json")
-
-	// Make the HTTP request
-	client := &http.Client{}
-	resp, err := client.Do(req)
-	if err != nil {
-		return nil, fmt.Errorf("failed to make API request: %w", err)
-	}
-	defer func() {
-		if err := resp.Body.Close(); err != nil {
-			backend.Logger.Warn("Failed to close response body", "error", err)
-		}
-	}()
-
-	// Check response status
-	if resp.StatusCode != http.StatusOK {
-		errorBody, _ := io.ReadAll(resp.Body)
-		return nil, fmt.Errorf("API request failed with status %d: %s", resp.StatusCode, string(errorBody))
-	}
-
-	// Read response body
-	body, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return nil, fmt.Errorf("failed to read response body: %w", err)
-	}
-
-	// Parse the API response - Cube returns an object with files array
-	var cubeFilesResponse struct {
-		Files []struct {
-			FileName string `json:"fileName"`
-			Content  string `json:"content"`
-		} `json:"files"`
-	}
-	if err := json.Unmarshal(body, &cubeFilesResponse); err != nil {
-		return nil, fmt.Errorf("failed to parse API response: %w", err)
-	}
-
-	// Convert to our response format
-	modelFiles := make([]ModelFile, len(cubeFilesResponse.Files))
-	for i, file := range cubeFilesResponse.Files {
-		modelFiles[i] = ModelFile{
-			FileName: file.FileName,
-			Content:  file.Content,
-		}
-	}
-
-	return &ModelFilesResponse{
-		Files: modelFiles,
-	}, nil
-}
-
-// DbSchemaResponse represents the response for the db-schema endpoint
-type DbSchemaResponse struct {
-	TablesSchema map[string]interface{} `json:"tablesSchema"`
-}
-
-// GenerateSchemaRequest represents the request for the generate-schema endpoint
-type GenerateSchemaRequest struct {
-	Format       string                 `json:"format"`
-	Tables       [][]string             `json:"tables"`
-	TablesSchema map[string]interface{} `json:"tablesSchema"`
-}
-
-// GenerateSchemaResponse represents the response for the generate-schema endpoint
-type GenerateSchemaResponse struct {
-	Files []GeneratedSchemaFile `json:"files"`
-}
-
-// GeneratedSchemaFile represents a generated schema file
-type GeneratedSchemaFile struct {
-	FileName string `json:"fileName"`
-	Content  string `json:"content"`
-}
-
-// handleDbSchema fetches database schema information from the Cube API
-func (d *Datasource) handleDbSchema(ctx context.Context, req *backend.CallResourceRequest, sender backend.CallResourceResponseSender) error {
-	// Fetch database schema from Cube API
-	dbSchema, err := d.fetchCubeDbSchema(ctx, req.PluginContext)
-	if err != nil {
-		backend.Logger.Error("Failed to fetch cube database schema", "error", err)
-		return sender.Send(jsonErrorResponse(500, errors.New("failed to fetch database schema from Cube API")))
-	}
-
-	// Marshal response
-	body, err := json.Marshal(dbSchema)
-	if err != nil {
-		backend.Logger.Error("Failed to marshal database schema response", "error", err)
-		return sender.Send(jsonErrorResponse(500, errors.New("failed to marshal response")))
-	}
-
-	return sender.Send(&backend.CallResourceResponse{
-		Status: 200,
-		Body:   body,
-		Headers: map[string][]string{
-			"Content-Type": {"application/json"},
-		},
-	})
-}
-
-// fetchCubeDbSchema fetches database schema from Cube's /playground/db-schema endpoint
-func (d *Datasource) fetchCubeDbSchema(ctx context.Context, pluginContext backend.PluginContext) (*DbSchemaResponse, error) {
-	// Build base URL and load configuration
-	apiReq, err := d.buildAPIURL(pluginContext, "")
-	if err != nil {
-		return nil, err
-	}
-
-	// Get base URL with test override support
-	baseURL := apiReq.Config.CubeApiUrl
-	if d.BaseURL != "" {
-		// Override for testing
-		baseURL = d.BaseURL
-	}
-
-	// Construct playground db-schema URL
-	baseURL = strings.TrimRight(baseURL, "/")
-	dbSchemaURL := baseURL + "/playground/db-schema"
-
-	// Create HTTP request
-	req, err := http.NewRequestWithContext(ctx, "GET", dbSchemaURL, nil)
-	if err != nil {
-		return nil, fmt.Errorf("failed to create request: %w", err)
-	}
-
-	if err := d.addAuthHeaders(req, apiReq.Config); err != nil {
-		return nil, fmt.Errorf("failed to add auth headers: %w", err)
-	}
-	req.Header.Set("Content-Type", "application/json")
-
-	// Make the HTTP request
-	client := &http.Client{}
-	resp, err := client.Do(req)
-	if err != nil {
-		return nil, fmt.Errorf("failed to make API request: %w", err)
-	}
-	defer func() {
-		if err := resp.Body.Close(); err != nil {
-			backend.Logger.Warn("Failed to close response body", "error", err)
-		}
-	}()
-
-	// Check response status
-	if resp.StatusCode != http.StatusOK {
-		errorBody, _ := io.ReadAll(resp.Body)
-		return nil, fmt.Errorf("API request failed with status %d: %s", resp.StatusCode, string(errorBody))
-	}
-
-	// Read response body
-	body, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return nil, fmt.Errorf("failed to read response body: %w", err)
-	}
-
-	// Parse the API response - Cube returns { tablesSchema: <schema_data> }
-	var cubeDbSchemaResponse struct {
-		TablesSchema map[string]interface{} `json:"tablesSchema"`
-	}
-	if err := json.Unmarshal(body, &cubeDbSchemaResponse); err != nil {
-		return nil, fmt.Errorf("failed to parse API response: %w", err)
-	}
-
-	return &DbSchemaResponse{
-		TablesSchema: cubeDbSchemaResponse.TablesSchema,
-	}, nil
-}
-
-// handleGenerateSchema generates Cube schema files from database schema
-func (d *Datasource) handleGenerateSchema(ctx context.Context, req *backend.CallResourceRequest, sender backend.CallResourceResponseSender) error {
-	// Only allow POST requests
-	if req.Method != "POST" {
-		return sender.Send(jsonErrorResponse(405, errors.New("method not allowed")))
-	}
-
-	// Parse request body
-	var generateSchemaReq GenerateSchemaRequest
-	if err := json.Unmarshal(req.Body, &generateSchemaReq); err != nil {
-		backend.Logger.Error("Failed to parse generate schema request", "error", err)
-		return sender.Send(jsonErrorResponse(400, errors.New("invalid request body")))
-	}
-
-	// Generate schema using Cube API
-	schemaResponse, err := d.fetchCubeGenerateSchema(ctx, req.PluginContext, &generateSchemaReq)
-	if err != nil {
-		backend.Logger.Error("Failed to generate cube schema", "error", err)
-		return sender.Send(jsonErrorResponse(500, errors.New("failed to generate schema from Cube API")))
-	}
-
-	// Marshal response
-	body, err := json.Marshal(schemaResponse)
-	if err != nil {
-		backend.Logger.Error("Failed to marshal generate schema response", "error", err)
-		return sender.Send(jsonErrorResponse(500, errors.New("failed to marshal response")))
-	}
-
-	return sender.Send(&backend.CallResourceResponse{
-		Status: 200,
-		Body:   body,
-		Headers: map[string][]string{
-			"Content-Type": {"application/json"},
-		},
-	})
-}
-
-// fetchCubeGenerateSchema generates schema files from Cube's /playground/generate-schema endpoint
-func (d *Datasource) fetchCubeGenerateSchema(ctx context.Context, pluginContext backend.PluginContext, generateSchemaReq *GenerateSchemaRequest) (*GenerateSchemaResponse, error) {
-	// Build base URL and load configuration
-	apiReq, err := d.buildAPIURL(pluginContext, "")
-	if err != nil {
-		return nil, err
-	}
-
-	// Get base URL with test override support
-	baseURL := apiReq.Config.CubeApiUrl
-	if d.BaseURL != "" {
-		// Override for testing
-		baseURL = d.BaseURL
-	}
-
-	// Construct playground generate-schema URL
-	baseURL = strings.TrimRight(baseURL, "/")
-	generateSchemaURL := baseURL + "/playground/generate-schema"
-
-	// Marshal request body
-	requestBody, err := json.Marshal(generateSchemaReq)
-	if err != nil {
-		return nil, fmt.Errorf("failed to marshal request body: %w", err)
-	}
-
-	// Create HTTP request
-	req, err := http.NewRequestWithContext(ctx, "POST", generateSchemaURL, strings.NewReader(string(requestBody)))
-	if err != nil {
-		return nil, fmt.Errorf("failed to create request: %w", err)
-	}
-
-	if err := d.addAuthHeaders(req, apiReq.Config); err != nil {
-		return nil, fmt.Errorf("failed to add auth headers: %w", err)
-	}
-	req.Header.Set("Content-Type", "application/json")
-
-	// Make the HTTP request
-	client := &http.Client{}
-	resp, err := client.Do(req)
-	if err != nil {
-		return nil, fmt.Errorf("failed to make API request: %w", err)
-	}
-	defer func() {
-		if err := resp.Body.Close(); err != nil {
-			backend.Logger.Warn("Failed to close response body", "error", err)
-		}
-	}()
-
-	// Check response status
-	if resp.StatusCode != http.StatusOK {
-		errorBody, _ := io.ReadAll(resp.Body)
-		return nil, fmt.Errorf("API request failed with status %d: %s", resp.StatusCode, string(errorBody))
-	}
-
-	// Read response body
-	body, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return nil, fmt.Errorf("failed to read response body: %w", err)
-	}
-
-	// Parse the API response - Cube returns { files: [{ fileName: "...", content: "..." }] }
-	var cubeGenerateSchemaResponse struct {
-		Files []struct {
-			FileName string `json:"fileName"`
-			Content  string `json:"content"`
-		} `json:"files"`
-	}
-	if err := json.Unmarshal(body, &cubeGenerateSchemaResponse); err != nil {
-		return nil, fmt.Errorf("failed to parse API response: %w", err)
-	}
-
-	// Convert to our response format
-	generatedFiles := make([]GeneratedSchemaFile, len(cubeGenerateSchemaResponse.Files))
-	for i, file := range cubeGenerateSchemaResponse.Files {
-		generatedFiles[i] = GeneratedSchemaFile{
-			FileName: file.FileName,
-			Content:  file.Content,
-		}
-	}
-
-	return &GenerateSchemaResponse{
-		Files: generatedFiles,
-	}, nil
-}
